Add HTML based design mockups
I’ve embedded the exported design mockups into HTML pages so that it’s easy for people taking the test to see how the mockups should display in the browser. Hopefully this will make the final goal of the test clear.
Note: the mockups were designed at a 1024px screen width.